+++ sdk/core/azure-core/test/ut/azure_core_test_traits.hpp
