+++ sdk/core/azure-core/test/ut/client_options_test.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-#include "test_traits.hpp"
+#include "azure_core_test_traits.hpp"
 
 #include <azure/core/context.hpp>
 #include <azure/core/http/http.hpp>
